moved git operations to core
