+++ core/Cargo.toml
@@ -50,11 +50,13 @@ crossbeam-channel = "^0.5"
 crossbeam = "^0.8"
 
 ## File system
+subprocess = "^0.2"
 jwalk = "^0.8"
 walkdir = "^2.5"
 relative-path = { version = "^1.9", features = ["serde"] }
 path-absolutize = "^3.1"
 glob = "^0.3"
+which = "^7.0"
 
 ## Logging and errors
 thiserror = "^1.0"
