+++ lib/Cargo.toml
@@ -55,7 +55,6 @@ crossbeam-channel = "^0.5"
 crossbeam = "^0.8"
 
 ## File system
-subprocess = "^0.2"
 relative-path = { version = "^1.9", features = ["serde"] }
 path-absolutize = "^3.1"
 glob = "^0.3"
